data/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.8.0
paho-mqtt>=1.6.1
aiosqlite>=0.20.0
python-multipart>=0.0.9
websockets>=12.0
python-dotenv>=1.0.0
pyserial>=3.5
PyYAML>=6.0.1 